Add installation event and move common to a folder
