+++ Cargo.toml
@@ -21,6 +21,9 @@ sha2 = "0.9.5"
 url = "2.2.2"
 serde_json = "1.0.64"
 
+# TODO: REMOVE
+serde_path_to_error = "0.1"
+
 [dependencies.matrix-sdk]
 git = "https://github.com/matrix-org/matrix-rust-sdk"
 features = ["default"]
